requirements.txt
fastapi
uvicorn
pytest
pytest-cov
pymongo
python-dotenv
dnspython
http3
pyrebase4 
firebase-admin
mongomock